+++ hubdle-gradle-plugin/testFixtures/kotlin/com/javiersc/hubdle/project/fixtures/KotlinVersionRegex.kt
@@ -1,0 +1,6 @@
+package com.javiersc.hubdle.project.fixtures
+
+internal const val KotlinVersionRegexOrEmptyOrNull = """(^$?)|(null)|(null)|(\d+\.\d+\.\d+)"""
+
+internal const val KotlinDevVersionRegexOrEmptyOrNull =
+    """(^$?)|(null)|(\d+\.\d+\.\d+)(\-)(dev)(\-)(\d+)"""