+++ hubdle-gradle-plugin/testFixtures/kotlin/com/javiersc/hubdle/project/fixtures/ciEnv.kt
@@ -1,0 +1,3 @@
+package com.javiersc.hubdle.project.fixtures
+
+internal const val CiEnv = "CI"